test : 채팅방 상세 조회 실패케이스 통합테스트 구현
- id를 null로 요청할 때
- id를 음수값으로 요청할 때
- 유효하지 않은 멤버가 요청할 때
- 유효하지 않은 chatRoomId로 요청할 때
