+++ src/test/java/louie/hanse/shareplate/integration/chatRoom/ChatRoomSearchIntegrationTest.java
@@ -102,53 +102,4 @@ class ChatRoomSearchIntegrationTest extends InitIntegrationTest {
             .body("message", equalTo(EMPTY_CHATROOM_INFO.getMessage()));
     }
 
-    @Test
-    void 회원의_채팅방_메세지를_조회한다() {
-        String accessToken = jwtProvider.createAccessToken(2370842997L);
-
-        given(documentationSpec)
-            .contentType(ContentType.JSON)
-            .filter(document("chatRoom-detail-of-member"))
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-
-            .when()
-            .get("/chatrooms/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value())
-            .body("chatRoomMemberId", equalTo(1))
-            .body("share.id", equalTo(1))
-            .body("share.writerId", equalTo(2370842997L))
-            .body("share.thumbnailImageUrl", equalTo("https://share-plate-file-upload.s3.ap-northeast-2.amazonaws.com/test/%E1%84%8B%E1%85%B5%E1%84%86%E1%85%B5%E1%84%8C%E1%85%B51.jpeg"))
-            .body("share.title  ", equalTo("강남역에서 떡볶이 먹을 사람 모집합니다."))
-            .body("share.price", equalTo(10000))
-            .body("share.originalPrice", equalTo(30000))
-            .body("share.cancel", equalTo(false))
-            .body("share.currentRecruitment", equalTo(3))
-            .body("share.finalRecruitment", equalTo(3))
-            .body("chats[0].contents", equalTo("내용1"))
-            .body("chats[0].writer", equalTo("정현석"))
-            .body("chats[0].writerThumbnailImageUrl", equalTo("http://k.kakaocdn.net/dn/wtMIN/btrII2nrJAv/KWEi4dNNGqeBYjzr0KZGK1/img_110x110.jpg"))
-            .body("chats[0].writtenDateTime", equalTo("2022-07-03 16:00"))
-            .body("chats[0].writtenByMe", equalTo(true));
-    }
-
-//    @Test
-//    void 회원이_문의용_일대일_채팅방을_생성한다() {
-//        String accessToken = jwtProvider.createAccessToken(2398606895L);
-//
-//        given(documentationSpec)
-//            .contentType(ContentType.JSON)
-//            .filter(document("chatRoom-for-inquire-in-question"))
-//            .header(AUTHORIZATION, accessToken)
-//            .body(Collections.singletonMap("shareId", 2))
-//
-//            .when()
-//            .post("/chatrooms")
-//
-//            .then()
-//            .statusCode(HttpStatus.OK.value());
-//    }
-
 }
